endorse/requirements.txt
mecab-python3
